# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Prospector API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus];

export const ContactStatus = {
  new: "new",
  contacted: "contacted",
  interested: "interested",
  not_interested: "not_interested",
  closed: "closed",
} as const;

export interface Contact {
  id: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  latitude: number;
  longitude: number;
  status: ContactStatus;
  notes?: string | null;
  propertyType?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateContactInputStatus =
  (typeof CreateContactInputStatus)[keyof typeof CreateContactInputStatus];

export const CreateContactInputStatus = {
  new: "new",
  contacted: "contacted",
  interested: "interested",
  not_interested: "not_interested",
  closed: "closed",
} as const;

export interface CreateContactInput {
  name: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  latitude: number;
  longitude: number;
  status: CreateContactInputStatus;
  notes?: string | null;
  propertyType?: string | null;
}

export type UpdateContactInputStatus =
  (typeof UpdateContactInputStatus)[keyof typeof UpdateContactInputStatus];

export const UpdateContactInputStatus = {
  new: "new",
  contacted: "contacted",
  interested: "interested",
  not_interested: "not_interested",
  closed: "closed",
} as const;

export interface UpdateContactInput {
  name?: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  latitude?: number;
  longitude?: number;
  status?: UpdateContactInputStatus;
  notes?: string | null;
  propertyType?: string | null;
}

export type ActionActionType =
  (typeof ActionActionType)[keyof typeof ActionActionType];

export const ActionActionType = {
  call: "call",
  visit: "visit",
  email: "email",
  meeting: "meeting",
  other: "other",
} as const;

export interface Action {
  id: number;
  contactId: number;
  title: string;
  description?: string | null;
  dueDate?: string | null;
  completed: boolean;
  actionType: ActionActionType;
  createdAt: string;
  updatedAt: string;
}

export type CreateActionInputActionType =
  (typeof CreateActionInputActionType)[keyof typeof CreateActionInputActionType];

export const CreateActionInputActionType = {
  call: "call",
  visit: "visit",
  email: "email",
  meeting: "meeting",
  other: "other",
} as const;

export interface CreateActionInput {
  contactId: number;
  title: string;
  description?: string | null;
  dueDate?: string | null;
  actionType: CreateActionInputActionType;
}

export type UpdateActionInputActionType =
  (typeof UpdateActionInputActionType)[keyof typeof UpdateActionInputActionType];

export const UpdateActionInputActionType = {
  call: "call",
  visit: "visit",
  email: "email",
  meeting: "meeting",
  other: "other",
} as const;

export interface UpdateActionInput {
  title?: string;
  description?: string | null;
  dueDate?: string | null;
  completed?: boolean;
  actionType?: UpdateActionInputActionType;
}

export type ListActionsParams = {
  contactId?: number;
};