# lib/api-zod/src/generated/types/action.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { ActionActionType } from "./actionActionType";

export interface Action {
  id: number;
  contactId: number;
  title: string;
  description?: string | null;
  dueDate?: Date | null;
  completed: boolean;
  actionType: ActionActionType;
  createdAt: Date;
  updatedAt: Date;
}