# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactStatus } from "./contactStatus";

export interface Contact {
  id: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  latitude: number;
  longitude: number;
  status: ContactStatus;
  notes?: string | null;
  propertyType?: string | null;
  createdAt: Date;
  updatedAt: Date;
}