# lib/api-zod/src/generated/types/createActionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActionInputActionType } from "./createActionInputActionType";

export interface CreateActionInput {
  contactId: number;
  title: string;
  description?: string | null;
  dueDate?: Date | null;
  actionType: CreateActionInputActionType;
}