# lib/api-zod/src/generated/types/createContactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactInputStatus } from "./createContactInputStatus";

export interface CreateContactInput {
  name: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  latitude: number;
  longitude: number;
  status: CreateContactInputStatus;
  notes?: string | null;
  propertyType?: string | null;
}