# lib/api-zod/src/generated/types/updateActionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateActionInputActionType } from "./updateActionInputActionType";

export interface UpdateActionInput {
  title?: string;
  description?: string | null;
  dueDate?: Date | null;
  completed?: boolean;
  actionType?: UpdateActionInputActionType;
}