# lib/api-zod/src/generated/types/updateContactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Prospector API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateContactInputStatus } from "./updateContactInputStatus";

export interface UpdateContactInput {
  name?: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  latitude?: number;
  longitude?: number;
  status?: UpdateContactInputStatus;
  notes?: string | null;
  propertyType?: string | null;
}